Turn off form tampering protection
Not really necessary, plus this was incompatible with any form that uses
any field not generated by FormHelper, which at least applies to the
registration page with a reCAPTCHA field. There may be some sort of
workaround, but since this level of security isn't necessary for this
application, I'm just turning the whole thing off.